進捗の計算のロジックが重複していたためリファクタ
