+++ src/components/typing/Progress.tsx
@@ -1,14 +1,22 @@
 type ProgressProps = {
-  typedTextCount: number;
-  targetTextCount: number;
-  accuracy: number;
+  targetText: string;
+  typedText: string;
 };
 
 export default function ProgressDisplay({
-  typedTextCount,
-  targetTextCount,
-  accuracy,
+  targetText,
+  typedText,
 }: ProgressProps) {
+  const correctTypedTextCount = typedText
+    .split("")
+    .filter((char, index) => char === targetText[index]).length;
+  const targetTextCount = targetText.length;
+  const typedTextCount = typedText.length;
+  const accuracy = parseFloat(
+    (
+      (typedTextCount > 0 ? correctTypedTextCount / typedTextCount : 1) * 100
+    ).toFixed(1)
+  );
   return (
     <>
       <p>
